+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/CHANGELOG.md
@@ -5,6 +5,10 @@
 Changed
 * GenericAudioManager, now returns tuple values for all PlayClipAtPos and PlayClip
 * GenericParticleManager, now returns the said particle system for PlayEffectAtPosition
+* ETransformType, added flags attribute
+* TransformMotionIJob, now do flag checks instead of switch case
+* SimpleRotator, now uses TransformMotionIJob
+* SimpleFollower, now uses TransformMotionIJob
 
 ## [1.0.8] - 2026-04-22